requirements.txt
flask
flask-cors
torch
transformers
autoawq
bitsandbytes